frontend/requirements.txt
# Frontend deps
streamlit
requests
python-dotenv
